+++ metadata/dashboard/dashboard_list_test.go
@@ -418,7 +418,7 @@ func (m *MockVariantsStore) Set(key, value string) error {
 	return nil
 }
 
-func (m *MockVariantsStore) Get(key string, opts ...query.Query) (string, error) {
+func (m *MockVariantsStore) Get(key string) (string, error) {
 	return "", nil
 }
